replaced original script
