+++ src/Stratis.FederatedPeg.IntegrationTests/Utils/PowerShellScriptGeneratorAsTests.cs
@@ -31,6 +31,7 @@ namespace Stratis.FederatedPeg.Tests.Utils
             this.newLine = s => stringBuilder.AppendLine(s);
 
             SetFolderVariables();
+            CreateHelpers();
             CopyStratisChainFiles();
             CreatePoaKeyFiles();
             AddCommentedFederationDetails();
@@ -39,6 +40,7 @@ namespace Stratis.FederatedPeg.Tests.Utils
             SetConsoleColors();
             StartGatewayDs();
             StartChainsD();
+            ShowRunningNodes();
             EnableWallets();
 
             this.output.WriteLine(stringBuilder.ToString());
@@ -64,29 +66,44 @@ namespace Stratis.FederatedPeg.Tests.Utils
         private void StartChainsD()
         {
             this.newLine("# MainchainUser");
-            this.newLine("cd $path_to_stratisd");
-            this.newLine($"start-process cmd -ArgumentList \"/k color {this.consoleColors[5]} && dotnet run -testnet -port=36178 -apiport=38221 -agentprefix=mainuser -datadir=$root_datadir\\MainchainUser -addnode=13.70.81.5 -addnode=52.151.76.252 -whitelist=52.151.76.252 -gateway=1\"");
-            this.newLine("timeout $interval_time");
-            this.newLine(Environment.NewLine);
+            CallStartNode(
+                path: "$path_to_stratisd",
+                title: "MAIN Chain User",
+                color: this.consoleColors[5],
+                args: $"-testnet -port=36178 -apiport=38221 -agentprefix=mainuser -datadir=$root_datadir\\MainchainUser -addnode=13.70.81.5 -addnode=52.151.76.252 -whitelist=52.151.76.252 -gateway=1",
+                timeout: "$interval_time"
+            );
 
-            this.newLine("# SidechainUser");
-            this.newLine("cd $path_to_sidechaind");
-            this.newLine($"start-process cmd -ArgumentList \"/k color {this.consoleColors[4]} && dotnet run -regtest -port=26179 -apiport=38225 -agentprefix=sideuser -datadir=$root_datadir\\SidechainUser agentprefix=sc_user -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 0)} -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 1)} -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 2)}\"");
-            this.newLine("timeout $interval_time");
-            this.newLine(Environment.NewLine);
+            CallStartNode(
+                path: "$path_to_sidechaind",
+                title: "SIDE Chain User",
+                color: this.consoleColors[4],
+                args: $"-regtest -port=26179 -apiport=38225 -agentprefix=sideuser -datadir=$root_datadir\\SidechainUser agentprefix=sc_user -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 0)} -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 1)} -addnode=127.0.0.1:36{GetPortNumberSuffix(this.chains[1], 2)}",
+                timeout: "$interval_time"
+            );
         }
 
         private void StartGatewayDs()
         {
-            this.newLine("cd $path_to_federationgatewayd");
-            federationMemberIndexes.ForEach(i => {
-                this.newLine($"# Federation member {i} main and side");
-                this.newLine(
-                    $"start-process cmd -ArgumentList \"/k color {this.consoleColors[i + 1]} && dotnet run -mainchain -testnet -agentprefix=fed{i + 1}main -datadir=$root_datadir\\gateway{i + 1} -port=36{GetPortNumberSuffix(this.chains[0], i)} -apiport=38{GetPortNumberSuffix(this.chains[0], i)} -counterchainapiport=38{GetPortNumberSuffix(this.chains[1], i)} -federationips=$mainchain_federationips -redeemscript=\"\"$redeemscript\"\" -publickey=$gateway{i + 1}_public_key -mincoinmaturity=1 -mindepositconfirmations=1\"");
-                this.newLine("timeout $long_interval_time");
-                this.newLine(
-                    $"start-process cmd -ArgumentList \"/k color {this.consoleColors[i + 1]} && dotnet run -sidechain -regtest -agentprefix=fed{i + 1}side -datadir=$root_datadir\\gateway{i + 1} -port=36{GetPortNumberSuffix(this.chains[1], i)} -apiport=38{GetPortNumberSuffix(this.chains[1], i)} -counterchainapiport=38{GetPortNumberSuffix(this.chains[0], i)} -federationips=$sidechain_federationips -redeemscript=\"\"$redeemscript\"\" -publickey=$gateway{i + 1}_public_key -mincoinmaturity=1 -mindepositconfirmations=1 -txindex=1\"");
-                this.newLine("timeout $long_interval_time");
+            this.newLine("#Federation members");
+            federationMemberIndexes.ForEach(i =>
+            {
+                CallStartNode(
+                    path: "$path_to_federationgatewayd",
+                    title: $"Gateway{i + 1} MAIN Chain",
+                    color: this.consoleColors[i + 1],
+                    args: $"-mainchain -testnet -agentprefix=fed{i + 1}main -datadir=$root_datadir\\gateway{i + 1} -port=36{GetPortNumberSuffix(this.chains[0], i)} -apiport=38{GetPortNumberSuffix(this.chains[0], i)} -counterchainapiport=38{GetPortNumberSuffix(this.chains[1], i)} -federationips=$mainchain_federationips -redeemscript=\"\"$redeemscript\"\" -publickey=$gateway{i + 1}_public_key -mincoinmaturity=1 -mindepositconfirmations=1",
+                    timeout: "$long_interval_time"
+                );
+
+                CallStartNode(
+                    path: "$path_to_federationgatewayd",
+                    title: $"Gateway{i + 1} SIDE Chain",
+                    color: this.consoleColors[i + 1],
+                    args: $"-sidechain -regtest -agentprefix=fed{i + 1}side -datadir=$root_datadir\\gateway{i + 1} -port=36{GetPortNumberSuffix(this.chains[1], i)} -apiport=38{GetPortNumberSuffix(this.chains[1], i)} -counterchainapiport=38{GetPortNumberSuffix(this.chains[0], i)} -federationips=$sidechain_federationips -redeemscript=\"\"$redeemscript\"\" -publickey=$gateway{i + 1}_public_key -mincoinmaturity=1 -mindepositconfirmations=1 -txindex=1",
+                    timeout: "$long_interval_time"
+                );
+
                 this.newLine(Environment.NewLine);
             });
         }
@@ -221,6 +238,45 @@ namespace Stratis.FederatedPeg.Tests.Utils
             this.consoleColors =
                 new Dictionary<int, string>() { { 1, "0E" }, { 2, "0A" }, { 3, "09" }, { 4, "0C" }, { 5, "0D" }, };
             this.newLine(Environment.NewLine);
+        }
+
+
+
+        private void CallStartNode(string path, string title, string color, string args, string timeout)
+        {
+            this.newLine($@"Start-Node -Path {path} -WindowTitle ""{title}"" -ConsoleColor {color} -CmdArgs ""{args}"" -Timeout {timeout}");
+        }
+
+
+
+        private void CreateHelpers()
+        {
+
+            this.newLine(@"
+cls
+
+[System.Collections.ArrayList]$running_nodes = @()
+
+function Start-Node {
+    param( [string]$Path, [string]$WindowTitle, [string]$ConsoleColor, [string]$CmdArgs, [int]$Timeout )
+
+    cd $Path
+    start-process cmd -ArgumentList ""/k title ${WindowTitle} && color ${ConsoleColor} && dotnet run ${CmdArgs}""
+    $running_nodes.Add(""------------------${WindowTitle}------------------`n${CmdArgs}"")
+    timeout $Timeout
+}
+
+function Show-Running-Nodes {
+    $body = $running_nodes -replace ""`n"",""&echo."" -join ""&echo.&echo.""
+    start-process cmd ""/k title Running Nodes && echo ${body}""
+}
+"
+            );
+        }
+
+        private void ShowRunningNodes()
+        {
+            this.newLine("Show-Running-Nodes");
         }
     }
 }
